Turn Cthulhu and JET into weakdeps
Create CthulhuExt and JETExt extension packages
