+++ src/parcel_snoopi_deep.jl
@@ -6,8 +6,6 @@ using AbstractTrees
 using Core.Compiler.Timings: InferenceFrameInfo
 using SnoopCompileCore: InferenceTiming, InferenceTimingNode, inclusive, exclusive
 using Profile
-using Cthulhu
-using JET
 
 const InferenceNode = Union{InferenceFrameInfo,InferenceTiming,InferenceTimingNode}
 
@@ -890,83 +888,14 @@ end
 AbstractTrees.children(tinf::InferenceTimingNode) = tinf.children
 
 InteractiveUtils.edit(itrig::InferenceTrigger) = edit(Location(itrig.callerframes[end]))
-Cthulhu.descend(itrig::InferenceTrigger; kwargs...) = descend(callerinstance(itrig); kwargs...)
-Cthulhu.instance(itrig::InferenceTrigger) = MethodInstance(itrig.node)
-Cthulhu.method(itrig::InferenceTrigger) = Method(itrig.node)
-Cthulhu.specTypes(itrig::InferenceTrigger) = Cthulhu.specTypes(Cthulhu.instance(itrig))
-Cthulhu.backedges(itrig::InferenceTrigger) = (itrig.callerframes,)
-Cthulhu.nextnode(itrig::InferenceTrigger, edge) = (ret = callingframe(itrig); return isempty(ret.callerframes) ? nothing : ret)
 
-"""
-    report_callee(itrig::InferenceTrigger)
-
-Return the `JET.report_call` for the callee in `itrig`.
-"""
-report_callee(itrig::InferenceTrigger; jetconfigs...) = report_call(Cthulhu.specTypes(itrig); jetconfigs...)
-
-"""
-    report_caller(itrig::InferenceTrigger)
-
-Return the `JET.report_call` for the caller in `itrig`.
-"""
-report_caller(itrig::InferenceTrigger; jetconfigs...) = report_call(Cthulhu.specTypes(callerinstance(itrig)); jetconfigs...)
-
-"""
-    report_callees(itrigs)
-
-Filter `itrigs` for those with a non-passing `JET` report, returning the list of `itrig => report` pairs.
-
-# Examples
-
-```jldoctest jetfib; setup=(using SnoopCompile, JET), filter=[r"\\d direct children", r"[0-9]*\\.?[0-9]+([eE][-+]?[0-9]+)?/[0-9]*\\.?[0-9]+([eE][-+]?[0-9]+)?"]
-julia> fib(n::Integer) = n ≤ 2 ? n : fib(n-1) + fib(n-2);
-
-julia> function fib(str::String)
-           n = length(str)
-           return fib(m)    # error is here
-       end
-fib (generic function with 2 methods)
-
-julia> fib(::Dict) = 0; fib(::Vector) = 0;
-
-julia> list = [5, "hello"];
-
-julia> mapfib(list) = map(fib, list)
-mapfib (generic function with 1 method)
-
-julia> tinf = @snoopi_deep try mapfib(list) catch end
-InferenceTimingNode: 0.049825/0.071476 on Core.Compiler.Timings.ROOT() with 5 direct children
-
-julia> @report_call mapfib(list)
-No errors detected
-```
-
-JET did not catch the error because the call to `fib` is hidden behind runtime dispatch.
-However, when captured by `@snoopi_deep`, we get
-
-```jldoctest jetfib; filter=[r"@ .*", r"REPL\\[\\d+\\]|none"]
-julia> report_callees(inference_triggers(tinf))
-1-element Vector{Pair{InferenceTrigger, JET.JETCallResult{JET.JETAnalyzer{JET.BasicPass{typeof(JET.basic_function_filter)}}, Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}}}}:
- Inference triggered to call fib(::String) from iterate (./generator.jl:47) inlined into Base.collect_to!(::Vector{Int64}, ::Base.Generator{Vector{Any}, typeof(fib)}, ::Int64, ::Int64) (./array.jl:782) => ═════ 1 possible error found ═════
-┌ @ none:3 fib(m)
-│ variable `m` is not defined
-└──────────
-```
-"""
-function report_callees(itrigs; jetconfigs...)
-    function rr(itrig)
-        rpt = try
-            report_callee(itrig; jetconfigs...)
-        catch err
-            @warn "skipping $itrig due to report_callee error" exception=err
-            nothing
-        end
-        return itrig => rpt
-    end
-    hasreport((itrig, report)) = report !== nothing && !isempty(JET.get_reports(report))
-
-    return [itrigrpt for itrigrpt in map(rr, itrigs) if hasreport(itrigrpt)]
-end
+# Implemented in JETExt
+"To use `report_caller` do `using Cthulhu, JET`"
+function report_caller end
+"To use `report_callee` do `using Cthulhu, JET`"
+function report_callee end
+"To use `report_callees` do `using Cthulhu, JET`"
+function report_callees end
 
 filtermod(mod::Module, itrigs::AbstractVector{InferenceTrigger}) = filter(==(mod) ∘ callermodule, itrigs)
 
@@ -1046,7 +975,6 @@ end
 
 InteractiveUtils.edit(node::TriggerNode) = edit(node.itrig)
 Base.stacktrace(node::TriggerNode) = stacktrace(node.itrig)
-Cthulhu.ascend(node::TriggerNode) = ascend(node.itrig)
 
 ### tagged trigger lists
 # good for organizing a collection of related triggers
@@ -1387,7 +1315,6 @@ unspec(s::Suggestion) = s ∈ (UnspecCall, UnspecType, CalleeVariable)
 unspec(s::Suggested)  = any(unspec, s.categories)
 
 Base.stacktrace(s::Suggested) = stacktrace(s.itrig)
-Cthulhu.ascend(s::Suggested) = ascend(s.itrig)
 InteractiveUtils.edit(s::Suggested) = edit(s.itrig)
 
 """
@@ -1581,8 +1508,12 @@ function unwrapconst(@nospecialize(arg))
         return arg.val
     elseif isa(arg, Core.PartialStruct)
         return arg.typ
-    elseif isa(arg, Core.Compiler.MaybeUndef)
-        return arg.typ
+    else
+        @static if VERSION <= v"1.10"
+            if isa(arg, Core.Compiler.MaybeUndef)
+                return arg.typ
+            end
+        end
     end
     return arg
 end
@@ -1911,6 +1842,9 @@ function max_end_time(node::InferenceTimingNode, recursive::Bool=false, tmax=-on
     end
     return tmax
 end
+
+# See ext/JETExt.jl for implementation
+function report_callees end
 
 for IO in (IOContext{Base.TTY}, IOContext{IOBuffer}, IOBuffer)
     for T = (InferenceTimingNode, InferenceTrigger, Precompiles, MethodLoc, MethodTriggers, Location, LocationTriggers)
